started plots
Began visualising data using the variables from earlier commits.

A plot of corrected mass flow rate against correct spool speed has bee created, and work has begun on creating a plot of isentropic efficiency against mass flow rate.

A issue that arose during this process was the incorrect use of the variable 'massFlow' and the lack of conversion of data from spreadsheet table to an array. The latter was fixed, and the former will be addressed in future.

This session was ended without calculating T3_is which will also be addressed in future