created plot 6
The correct mass intake values were found using the equation provided in the coursework brief.
Several new values and constants were also found as they aided in the calculation of necessary variables and figures.
A function for finding the density at ay temperature and altitude was defined and used t calculate the intake air density.
A mistake in the values used to calculate the compressor corrected mass flow rate was rectified. The mass flow rate being used before this correction was obtained from the engine data. These values are wrong and the correct values were calculated using and equation provided in the coursework brief and stored in the variable "intakeMassFlow"