plots complete
The final plot assigned to me has been completed.

The exit mach number was calculated by assuming that the compressor blades were the same length as the intake diameter and, the compressor rotates at the same speed as the shaft. In addition, the compressor entry temperature was assumed to be the same as its exit and this value was used to calculate exit velocity.

The value for gas constant was changed to use the appropriate value. The previous one is used in chemistry so, it was changed to the value engineers would use.