finalised v1
All the plots make sense, the data is consistent with literature and after much consultation with Dr. Chandra, all the minor calculation missteps have been resolved.

This script is the final form of the data without styling of the plots, order of the script and other quality of life, aesthetic modifications.